🩹 Refactor: AuthenticationViewModel 기능 분리
로그인/가입 관련 데이터는 LoginViewModel에서 처리할 수 있도록 분리
