+++ PADO/PADO/iOS/Login/View/StartView.swift
@@ -8,18 +8,10 @@
 import SwiftUI
 
 struct StartView: View {
-    @EnvironmentObject var viewModel: AuthenticationViewModel
-    
-    @State private var currentIndex: Int = 0
-    @State private var titleText: [TextAnimation] = []
-    @State private var subTitleAnimation: Bool = false
-    @State private var endAnimation = false
+    @StateObject var loginVM = LoginViewModel()
 
     @Binding var isShowStartView: Bool
     
-    let titles = ["Clean your mind from", "Unique experience", "The ultimate sns"]
-    let subTitles = ["Decorate your friend's picture", "Prepare your mind for sweet dreams", "Healty mind - better think - well being"]
-
     var body: some View {
         NavigationStack {
             ZStack {
@@ -30,7 +22,7 @@ struct StartView: View {
                             .resizable()
                             .scaledToFill()
                             .frame(width: size.width, height: size.height)
-                            .opacity(currentIndex == (index - 1) ? 1 : 0)
+                            .opacity(loginVM.titleIndex == (index - 1) ? 1 : 0)
                     }
                     LinearGradient(colors: [.clear, .black.opacity(0.5), .black], startPoint: .top, endPoint: .bottom)
                 }
@@ -41,34 +33,36 @@ struct StartView: View {
                     
                     VStack {
                         HStack(spacing: 0) {
-                            ForEach(titleText) { text in
+                            ForEach(loginVM.titleText) { text in
                                 Text(text.text)
                                     .offset(y: text.offset)
                             }
                             .font(.largeTitle.bold())
                         }
-                        .offset(y: endAnimation ? -50 : 0)
-                        .opacity(endAnimation ? 0 : 1)
+                        .offset(y: loginVM.endAnimation ? -50 : 0)
+                        .opacity(loginVM.endAnimation ? 0 : 1)
 
-                        Text(subTitles[currentIndex])
+                        Text(loginVM.subTitles[loginVM.titleIndex])
                             .opacity(0.7)
-                            .offset(y: !subTitleAnimation ? 70 : 0)
-                            .offset(y: endAnimation ? -50 : 0)
-                            .opacity(endAnimation ? 0 : 1)
+                            .offset(y: !loginVM.subTitleAnimation ? 70 : 0)
+                            .offset(y: loginVM.endAnimation ? -50 : 0)
+                            .opacity(loginVM.endAnimation ? 0 : 1)
                             .padding(.top, 5)
                     }
                     .padding(.bottom, 80)
                     
                     HStack(spacing: 20) {
                         NavigationLink {
-                            SignUpView(loginSignUpType: LoginSignUpType.signUp,
+                            SignUpView(loginVM: loginVM,
+                                       loginSignUpType: LoginSignUpType.signUp,
                                        isShowStartView: $isShowStartView)
                         } label: {
                             SignUpButton(text: "회원가입")
                         }
                         
                         NavigationLink {
-                            SignUpView(loginSignUpType: LoginSignUpType.login,
+                            SignUpView(loginVM: loginVM,
+                                       loginSignUpType: LoginSignUpType.login,
                                        isShowStartView: $isShowStartView)
                         } label: {
                             EnjoyButton(text: "로그인")
@@ -79,57 +73,12 @@ struct StartView: View {
                 .frame(maxWidth: .infinity, maxHeight: .infinity, alignment: .bottom)
             }
             .onAppear {
-                if titleText.isEmpty {
-                    startAnimation()
+                if loginVM.titleText.isEmpty {
+                    loginVM.startAnimation()
                 }
             }
-            .onChange(of: currentIndex) { _, _ in
-                startAnimation()
-            }
-        }
-    }
-
-    func getSpilitedText(text: String, completion: @escaping () -> Void) {
-        for (index, character) in text.enumerated() {
-            titleText.append(TextAnimation(text: String(character)))
-
-            DispatchQueue.main.asyncAfter(deadline: .now() + Double(index) * 0.03) {
-                withAnimation(.easeInOut(duration: 0.45)) {
-                    titleText[index].offset = 0
-                }
-            }
-        }
-
-        DispatchQueue.main.asyncAfter(deadline: .now() + Double(text.count) * 0.02) {
-            withAnimation(.easeInOut(duration: 0.75)) {
-                subTitleAnimation.toggle()
-            }
-        }
-
-        DispatchQueue.main.asyncAfter(deadline: .now() + 0.95) {
-            completion()
-        }
-    }
-
-    func startAnimation() {
-        titleText.removeAll()
-        subTitleAnimation = false
-        endAnimation = false
-
-        getSpilitedText(text: titles[currentIndex]) {
-            withAnimation(.easeInOut(duration: 1)) {
-                endAnimation.toggle()
-            }
-
-            DispatchQueue.main.asyncAfter(deadline: .now() + 0.9) {
-                subTitleAnimation.toggle()
-                endAnimation.toggle()
-
-                if currentIndex < (titles.count - 1) {
-                    currentIndex += 1
-                } else {
-                    currentIndex = 0
-                }
+            .onChange(of: loginVM.titleIndex) { _, _ in
+                loginVM.startAnimation()
             }
         }
     }
